Give backdrop elements an ID
This allows them to be reattached correctly if the Modal instance was
removed but the backdrop element was not. This can happen due to
caching and page navigation (i.e. you open the modal, navigate away,
then back again - the cached page has the backdrop element, but the
modal is not visible).